2 week return type fix

+++ solutions/aleks_apalikov/Craft_2/sources/2_4_Deals/Deals.cpp
@@ -29,7 +29,7 @@ int Deals::createOutput()
 	}
 	int n = 0; //total number of objects
 	UINT32 type, time, currentTime = 0;
-	size_t len;
+	UINT32 len;
 	char* str = NULL;
 	bool toWrite = false;
 	DealsElem* de = NULL;
